sample_apps/simple_api/api/src/requirements.txt
Werkzeug
SQLAlchemy==1.1.9
Flask_Caching
Flask_SQLAlchemy==2.2
click
marshmallow
Flask_Bcrypt
flask_apispec
Flask
PyJWT
Flask-JWT-Extended
unicode_slugify
psycopg2
Flask-Migrate
gunicorn
Flask-Cors